+++ src/drivers/broadcast-channel.ts
@@ -48,7 +48,7 @@ export class BroadcastChannelDriver extends Driver {
     this.#emitter.off(event, handler);
   }
 
-  async dispatch(namespace: string[], data: number[]) {
+  async publish(namespace: string[], data: number[]) {
     const [event] = namespace.slice(-1);
     this.#bc.postMessage([event, data]);
   }